redirect flow added in react example
