+++ examples/react-app/src/App.css
@@ -1,4 +1,46 @@
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
+}
+
 .App {
+  font-family: 'Avenir', Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
   text-align: center;
-  padding-top: 60px;
+  color: #2c3e50;
+  margin-top: 60px;
+}
+#console {
+  border: 1px solid black;
+  height: 40px;
+  padding: 2px;
+  bottom: 10px;
+  position: absolute;
+  text-align: left;
+  width: calc(100% - 20px);
+  border-radius: 5px;
+}
+#console::before {
+  content: 'Console :';
+  position: absolute;
+  top: -20px;
+  font-size: 12px;
+}
+#console > p {
+  margin: 0.5em;
+}
+button {
+  height: 25px;
+  margin: 5px;
+  background: none;
+  border-radius: 5px;
+  cursor: pointer;
 }
