+++ examples/react-app/src/index.css
@@ -1,4 +1,4 @@
-body {
+/* body {
   margin: 10px;
   font-family: Avenir, Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
@@ -26,4 +26,4 @@ body {
 .console > p {
   margin: 0.5em;
   word-wrap: break-word;
-}
+} */
