+++ examples/react-app/src/index.tsx
@@ -1,11 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
+
 import './index.css';
-import App from './App';
+import RouterPage from './router';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Router>
+
+      <RouterPage />
+    </Router>
   </React.StrictMode>,
   document.getElementById('root'),
 );
